vape_bot/requirements.txt
aiogram==3.3.0
python-dotenv==1.0.0
aiohttp==3.9.1